requirements.txt
asgiref==3.6.0
Django==4.2
django-appconf==1.0.5
django-ckeditor==6.5.1
django-environ==0.10.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-select2==8.1.1
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.5.0
